create a PlayerMatchesTable component
